[ui][settings] refactoring
